requirements.txt
Pillow>=9.2.0
pyzbar>=0.1.9
qrcode>=7.3.1
